refactor: migrate browser observation to core
